julia version for those who wish to have it.
I know that the project has "python" in it's name but I really liked the example in spite of my  tend to preference of julia. This is just a bare-bones copy without the comments or any attempt to optimize. In spite of this it runs over twice as fast as the python version on my machine. I don't know how fair the comparison is. There is one small difference... I did not seed the random number generator... just out of laziness.