Pollish code
In keeping this as much like the original example as possible I went ahead and seeded the rand function. On the other hand, I used randn (normalised) instead since that is something that's built in in julia. I also took code out of global scope since that is highly discouraged in julia due to the speed penalty imposed.